feat: add endpoint to list books
