+++ src/backend/Cargo.toml
@@ -13,6 +13,8 @@ serde_json = "1.0"
 jsonwebtoken = "8"
 derive_more = "^0.99"
 actix-cors = "0.6.2"
+reqwest = { version = "0.11", features = ["json"] }
+async-trait = "0.1.57"
 
 [dependencies.uuid]
 version = "1.1.2"
