+++ src/backend/src/adapters/mod.rs
@@ -1,3 +1,4 @@
+pub mod book_repository;
 pub mod generate_jwt;
 pub mod password_hash;
 pub mod user_repository;
